fix: websocket in dashboard
