+++ medecin_mobile/lib/services/web_socket_services.dart
@@ -126,21 +126,11 @@ class WebSocketService {
     _channel?.sink.add(readMessage);
   }
 
-  void askMobileConnection(String uuid, String email, String password) {
-    final askMobileConnection = jsonEncode({
-      'action': 'askMobileConnection',
-      'uuid': uuid,
-      'type': 1,
-      'email': email,
-      'password': password,
-    });
-    _channel?.sink.add(askMobileConnection);
-  }
-
   void responseMobileConnection(String patientAuthTokenWS, String uuid) {
+    Logger().i('ResponseMobileConnection: $patientAuthTokenWS, $uuid');
     final responseMobileConnection = jsonEncode({
-      'action': 'responseMobileConnection',
-      'patientAuthTokenWS': patientAuthTokenWS,
+      'action': 'responseMobileConnection ',
+      'authToken': patientAuthTokenWS,
       'uuid': uuid,
       "response": true,
     });
@@ -160,10 +150,10 @@ class WebSocketService {
       case 'receive_message':
         onReceiveMessage?.call(decodedMessage);
         break;
-      case 'askMobileConnection':
+      case 'ask_mobile_connection':
         onAskMobileConnection?.call(decodedMessage);
         break;
-      case 'responseMobileConnection':
+      case 'response_mobile_connection':
         onResponseMobileConnection?.call(decodedMessage);
         break;
       case 'read_message':
